# docs/ai/visual-operation-guard/integration/visual-operation-guard.openapi.yaml
openapi: 3.1.0
info:
  title: AIEfficiency Visual Operation Guard API
  version: 1.0.0
  description: >-
    Structured-first UI automation authorization, one-time execution, result,
    checkpoint, and human-handoff service. Tool calls use the sequence
    authorize -> begin -> result.
servers:
- url: http://127.0.0.1:18081
security:
- bearerAuth: []
paths:
...
  /v1/tasks:
    post:
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '409':
          description: The taskId already exists; task budgets are never reset by duplicate create
  /v1/tasks/{taskId}:
    parameters:
    - $ref: '#/components/parameters/TaskId'
    get:
      operationId: getTask
      responses:
        '200':
          description: Task state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  schemas:
    CreateTaskRequest:
      type: object
      properties:
        taskId:
          type: string
        mode:
          type: string
          default: balanced
        taskType:
          type: string
          default: generic
        unattended:
          type: boolean
          default: false
...
    TaskResponse:
      type: object
      required:
      - taskId
      - state
      properties:
        taskId:
          type: string
        state:
          type: object
          additionalProperties: true